+++ components/Footer/Footer.jsx
@@ -1,0 +1,11 @@
+import styles from './Footer.module.css'
+
+export default function Footer() {
+  return (
+    <>
+      <footer className={styles.footer}>
+        <img src="/Argo.svg" alt="Argo Logo" className={styles.logo} />
+      </footer>
+    </>
+  )
+}
