+++ components/Header/Header.jsx
@@ -1,0 +1,13 @@
+import styles from './Header.module.css'
+
+export default function Footer() {
+  return (
+    <>
+      <header className={styles.footer}>
+        <h1>
+          My website
+        </h1>
+      </header>
+    </>
+  )
+}
