+++ package.json
@@ -8,9 +8,12 @@
     "export": "next export"
   },
   "dependencies": {
-    "cypress": "^10.0.3",
+    "cypress": "^4.2.0",
     "next": "^12.1.5",
     "react": "^18.0.0",
     "react-dom": "^18.0.0"
+  },
+  "devDependencies": {
+    "sass": "^1.64.1"
   }
 }
