--- pages/_app.js
@@ -1,7 +1,0 @@
-import '@styles/globals.css'
-
-function Application({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}
-
-export default Application
