+++ pages/_app.jsx
@@ -1,0 +1,7 @@
+import '@styles/globals.css'
+
+function Application({ Component, pageProps }) {
+  return <Component {...pageProps} />
+}
+
+export default Application
