+++ pages/index.jsx
@@ -1,0 +1,7 @@
+import Accueil from "./accueil"
+
+export default function Home() {
+  return (
+    <Accueil/>
+  )
+}
